backend/requirements.txt
# python==3.12
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
oss2==2.18.0
httpx==0.27.0
